+++ src/types.ts
@@ -1,3 +1,5 @@
+import { LegendDisplayMode, LegendPlacement } from '@grafana/ui';
+
 type SortBy = 'text' | 'startTime';
 type SortOrder = 'asc' | 'desc';
 
@@ -9,6 +11,10 @@ export interface GanttOptions {
   labelFields: string[];
   sortBy: SortBy;
   sortOrder: SortOrder;
+
+  colors: Record<string, string>;
+  legendMode: LegendDisplayMode;
+  legendPlacement: LegendPlacement;
 
   experiments: {
     enabled: boolean;
